refactor: discord embeds, fix book covers
- move embed creation to shared module
- enhance image URL handling
- both slash commands and chatbot use same embed logic

Fixes #12
